+++ handshake-nodejs-reporters/packages/handshake-dashboard/next.config.js
@@ -4,7 +4,7 @@ const nextConfig = {
   output: 'export',
   // Optional: Change links `/me` -> `/me/` and emit `/me.html` -> `/me/index.html`
   trailingSlash: true,
-  transpilePackages: ['handshake', 'echarts', 'zrender'],
+  transpilePackages: ['handshake-dashboard', 'echarts', 'zrender'],
   env: {
     NEXT_PY_PORT: process.env.NEXT_PY_PORT ?? '6969'
   },
@@ -14,4 +14,5 @@ const nextConfig = {
   distDir: process.env.EXPORT_DIR ?? "dist"
 }
 
+// eslint-disable-next-line unicorn/prefer-module
 module.exports = nextConfig