+++ handshake-nodejs-reporters/packages/wdio-handshake-reporter/CHANGELOG.md
@@ -1,5 +1,20 @@
 # wdio-graspit-reporter
 
+## 0.3.5
+
+### Patch Changes
+
+- 28bea4c: feat: add options for the number of handshake-server workers
+- Updated dependencies [28bea4c]
+  - common-handshakes@0.3.5
+
+## 0.3.4
+
+### Patch Changes
+
+- Updated dependencies
+  - common-handshakes@0.3.4
+
 ## 0.3.2
 
 ### Patch Changes